% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatedRightAwayAndDeath.R
\docType{package}
\name{TreatedRightAwayAndDeath}
\alias{TreatedRightAwayAndDeath}
\alias{TreatedRightAwayAndDeath-package}
\title{TreatedRightAwayAndDeath}
\description{
TreatedRightAwayAndDeath
}
